refactor: extract CreateSafeCard component from CreateSafePage for reusability
